+++ src/Feature/Sms/Sendernames/SendernamesFeature.php
@@ -7,7 +7,6 @@ use Smsapi\Client\Feature\Sms\Sendernames\Bag\ActivateSendernameBag;
 use Smsapi\Client\Feature\Sms\Sendernames\Bag\CreateSendernameBag;
 use Smsapi\Client\Feature\Sms\Sendernames\Bag\DeleteSendernameBag;
 use Smsapi\Client\Feature\Sms\Sendernames\Bag\FindSendernameBag;
-use Smsapi\Client\Feature\Sms\Sendernames\Bag\FindSendernamesBag;
 use Smsapi\Client\Feature\Sms\Sendernames\Bag\MakeSendernameDefaultBag;
 use Smsapi\Client\Feature\Sms\Sendernames\Data\Sendername;
 
@@ -18,7 +17,7 @@ interface SendernamesFeature
 {
     public function findSendername(FindSendernameBag $findSendernameBag): Sendername;
 
-    public function findSendernames(FindSendernamesBag $getSendernamesBag = null): array;
+    public function findSendernames(): array;
 
     public function createSendername(CreateSendernameBag $createSendernameBag): Sendername;